Fix Windows build failures and add 4 Windows arch targets
- Fix guide_codex.hpp missing cstdint causing MSVC/MinGW error: enum class GuideCategory : uint8_t needs cstdint
- Add cstdint include to all headers using uint8_t/uint32_t: config, monster, player, session, npc, shop, dungeon
- Fix ascii_art.cpp Wconversion warnings: innerW * pct and h * pct cast to float before multiply
- Release workflow: expand Windows matrix from 2 to 4 archs
  - windows-x86_64 (existing)
  - windows-x86 (32-bit) artifact shadowdeep-Zv1-windows-x86.zip
  - windows-arm64 (existing)
  - windows-arm32 (32-bit) artifact shadowdeep-Zv1-windows-arm32.zip
- UpdateManager: getArch now detects x86 32-bit (_M_IX86, __i386__) returning x86, and arm32 returning armv7
- getCandidateArtifactNames: Windows now includes x86, x32, i386, arm32, arm, armv7 fallbacks plus all 4 main artifacts
- Release body updated to list 4 Windows artifacts
- Verified g++ -std=c++20 with -Wconversion builds without ascii_art warnings
- No emojis, only colored ASCII letters maintained

+++ src/tui/ascii_art.cpp
@@ -679,7 +679,7 @@ void drawProgressBar(ScreenBuffer& screen, int x, int y, int w, int current, int
     if (pct < 0) pct = 0;
     if (pct > 1) pct = 1;
     int innerW = w - 2;
-    int filled = static_cast<int>(innerW * pct);
+    int filled = static_cast<int>(static_cast<float>(innerW) * pct);
     for (int i = 0; i < innerW; ++i) {
         if (i < filled) screen.set(x + 1 + i, y + 1, "#", fill, true);
         else screen.set(x + 1 + i, y + 1, "-", empty, false);
@@ -695,7 +695,7 @@ void drawVerticalBar(ScreenBuffer& screen, int x, int y, int h, int current, int
     float pct = static_cast<float>(current) / static_cast<float>(max);
     if (pct < 0) pct = 0;
     if (pct > 1) pct = 1;
-    int filled = static_cast<int>(h * pct);
+    int filled = static_cast<int>(static_cast<float>(h) * pct);
     for (int i = 0; i < h; ++i) {
         if (i < h - filled) screen.set(x, y + i, ".", empty, false);
         else screen.set(x, y + i, "#", fill, true);
